Switch to use ParsedResult for APIHandler.
