+++ packages/board-server/src/server/types.ts
@@ -6,8 +6,38 @@
 
 import type { IncomingMessage, ServerResponse } from "http";
 
-export type ApiHandler = (
-  path: string,
+export type GeneralRequestType = "list" | "create";
+
+export type UserRequestType =
+  | "list"
+  | "create"
+  | "get"
+  | "update"
+  | "app"
+  | "api"
+  | "invoke"
+  | "describe";
+
+export type RequestType = GeneralRequestType | UserRequestType;
+
+export type GeneralParseResult = {
+  success: true;
+  type: GeneralRequestType;
+};
+
+export type BoardParseResult = {
+  success: true;
+  type: UserRequestType;
+  board: string;
+};
+
+export type ParseResult =
+  | GeneralParseResult
+  | BoardParseResult
+  | { success: false; error: string; code: number };
+
+export type ApiHandler = <T extends ParseResult>(
+  parsed: T,
   req: IncomingMessage,
   res: ServerResponse,
   body?: unknown
